Refactoring

+++ src/client-handlers.js
@@ -1,38 +1,68 @@
+const path = require('path');
+const client = global.client;
 const userIds = {
-    PiercingGoblin: '<@187698900843495424>'
+    PiercingGoblin: '179307528079802369',
+    Gamervixen89: '113071920030679040'
 };
 
 function babyBoy(msg) {
     if (msg.author.id !== '179307528079802369') {
-        msg.reply(`${userIds.PiercingGoblin} is a baby boy!`);
+        msg.reply(`<@${userIds.PiercingGoblin}> is a baby boy!`);
     } else {
         msg.reply('you are a baby boy!');
     }
 }
 
-function deleteMessage(msg) {
-    msg.delete();
+function makeBotInvite(msg) {
+    client.generateInvite(['ADMINISTRATOR']).then((inv) => {
+        msg.reply(inv);
+    });
 }
 
-function join(msg) {
-    if (msg.member.voiceChannel) {
-        msg.member.voiceChannel.join()
-            .then(connection => { // Connection is an instance of VoiceConnection
-                const dispatcher = connection.playFile('../media/audio/vlad.ogx');
+function playFile(msg, fileName, volume = 1, deleteSourceMessage = true, channelToJoin) {
+    if (deleteSourceMessage) {
+        msg.delete()
+            .catch((err) => {
+                if (err.code === 50013) {
+                    console.log('Failed to delete message. Check bot permissions');
+                } else {
+                    console.log(err);
+                }
+            });
+    }
+
+    let voiceChannel;
+    if (channelToJoin) {
+        voiceChannel = client.channels.get(channelToJoin);
+        if (!voiceChannel) {
+            msg.member.createDM().then((dm) => {
+                dm.send(`You specified an invalid channel ID! ${channelToJoin} is not a valid channel!`);
+            });
+        }
+    } else if (msg.member && msg.member.voiceChannel) {
+        voiceChannel = msg.member.voiceChannel;
+        if (!voiceChannel) {
+            msg.member.createDM().then((dm) => {
+                dm.send('You must be in a voice hannel in order to use voice commands!');
+            });
+        }
+    }
+
+    if (voiceChannel) {
+        voiceChannel.join()
+            .then(connection => {
+                const dispatcher = connection.playFile(path.resolve(`./media/audio/${fileName}`));
+                dispatcher.setVolume(volume);
                 dispatcher.on('end', () => {
-                    msg.member.voiceChannel.leave();
+                    voiceChannel.leave();
                 });
             })
             .catch(console.log);
-    } else {
-        msg.author.createDM().then((dm) => {
-            dm.sendMessage('You must be in a voice channel!');
-        });
     }
 }
 
 module.exports = {
     babyBoy,
-    deleteMessage,
-    join
+    makeBotInvite,
+    playFile
 };
